+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Lobby / Waiting room move button / password (!852)
 - Add unread indicator for chat messages (!844)
 - Add legal vote token copy field. (!836)
+- Lobby without password label as invited (!863)
 
 ## 1.2.0